+++ src/main/java/br/com/desafio/repository/IvotoDAO.java
@@ -6,16 +6,17 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
 public interface IvotoDAO extends JpaRepository<Voto, Long> {
 
     //Spring Data com Query methods
-    Optional<Voto> findByCodPautaAndCpf(Long codPauta, String cpf);
-    List<Voto> findByCodPauta(Long codPauta, Pageable pageable);
+    Optional<Voto> findByCodPautaAndCpf(BigInteger codPauta, String cpf);
+    List<Voto> findByCodPauta(BigInteger codPauta, Pageable pageable);
 
     //Spring Data com @Query e @NamedNativeQuery @SqlResultSetMapping na entidade Voto
     @Query(nativeQuery = true)
-    List<ApuracaoDTO> GetApuracao(Long codPauta);
+    List<ApuracaoDTO> GetApuracao(BigInteger codPauta);
 }
