+++ src/main/java/br/com/desafio/service/IvotoService.java
@@ -4,11 +4,11 @@ import br.com.desafio.domain.entity.Voto;
 import br.com.desafio.model.VotoDTO;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface IvotoService {
 
      Voto create(VotoDTO obj);
-     List<Voto> findByCodPauta(Long codPauta, Pageable pageable);
-     //List<ApuracaoDTO> apuracao(Long codPauta);
+     List<Voto> findByCodPauta(BigInteger codPauta, Pageable pageable);
 }
